更改 decimal

+++ src/main/java/org/example/calculator/core/parser/nodes/ASTNode.java
@@ -1,5 +1,7 @@
 package org.example.calculator.core.parser.nodes;
 
+import java.math.BigDecimal;
+
 /**
  * 语法树节点接口
  */
@@ -9,5 +11,5 @@ public interface ASTNode {
      *
      * @return 返回节点的计算值
      */
-    double evaluate();
+    BigDecimal evaluate();
 }
